requirements.txt
# --- Framework + typing ---
fastapi==0.110.0
starlette==0.36.3
pydantic==1.10.13
typing_extensions==4.9.0

# --- LangChain stack ---
langchain==0.2.11
langchain-core==0.2.32
langserve==0.2.1
langchain-openai==0.1.20
langchain-ollama==0.1.0

# --- Utils ---
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.30.1

